requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
boto3==1.21.23
botocore==1.24.23
certifi==2022.5.18.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Django==3.2.12
django-cors-headers==3.11.0
django-graphql-jwt==0.3.4
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.13.1
Fiona==1.8.21
geopandas==0.10.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
jmespath==1.0.0
munch==2.5.0
numpy==1.22.4
pandas==1.4.2
Pillow==9.0.1
promise==2.3
psycopg2==2.9.3
PyJWT==2.3.0
pyproj==3.3.1
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2021.3
Rx==1.6.1
s3transfer==0.5.2
sendgrid==6.9.7
Shapely==1.8.2
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.9
